% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_abonnement.R
\name{abonnement_server}
\alias{abonnement_server}
\title{Shiny module providing server logic for the subscription tab}
\usage{
abonnement_server(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\description{
Shiny module providing server logic for the subscription tab
}
