% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataOps.R
\name{basicSample}
\alias{basicSample}
\title{Sample the data}
\usage{
basicSample(data, gender, status, time1, time2, age1, age2, unit)
}
\description{
This function extracts samples from the large dataset. This function works
on datasets that have been preprosessed by prePros.
Sorting on gender, status, date of treatment measuring point, age of patient and unit.
}
