% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegData.R
\name{getTimepointData}
\alias{getTimepointData}
\title{Provides a dataframe containing data from a registry}
\usage{
getTimepointData()
}
\value{
regData data frame
}
\description{
Provides a dataframe containing data from a registry
}
