% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQualind.R
\name{kval_plot}
\alias{kval_plot}
\title{Create a Quality Indicator Plot}
\usage{
kval_plot(data, ggData, anno)
}
\arguments{
\item{data}{A data frame or data object containing the quality indicator data to be plotted.}

\item{ggData}{A data frame formatted for use with ggplot2, containing processed data for visualization.}

\item{anno}{A list or data frame containing annotation information to be added to the plot.}
}
\value{
A ggplot2 plot object representing the quality indicator visualization.
}
\description{
Generates a quality indicator plot based on the provided data and graphical parameters.
}
\keyword{internal}
