% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHist.R
\name{makeHist}
\alias{makeHist}
\title{Make a histogram}
\usage{
makeHist(df, var)
}
\arguments{
\item{df}{dataframe from which output is to be made}

\item{var}{string defining which varable in the data frame to use}
}
\value{
a graphical object
}
\description{
Make a histogram
}
\examples{
makeHist(df = mtcars, var = "mpg")
}
