% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_qualind.R
\name{module_kvalitetsindikator_ui}
\alias{module_kvalitetsindikator_ui}
\title{UI Module for Kvalitetsindikator}
\usage{
module_kvalitetsindikator_ui(id)
}
\arguments{
\item{id}{A unique identifier for the module namespace.}
}
\value{
A Shiny UI element for the module.
}
\description{
This function defines the UI for the "Kvalitetsindikator" module.
}
\details{
This module is designed to create the user interface for the
"Kvalitetsindikator" functionality within the application. The `id` parameter
is used to create a namespace for the module, ensuring that its UI elements
are uniquely identified within the Shiny application.
}
