% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_qualind.R
\name{module_qualind_server}
\alias{module_qualind_server}
\title{Server Module for Quality Indicators}
\usage{
module_qualind_server(id)
}
\arguments{
\item{id}{A unique identifier for the module namespace.}
}
\value{
A Shiny server function for the Quality Indicators module.
}
\description{
This function defines the server logic for the Quality Indicators module.
}
\details{
This module is designed to handle the server-side logic for displaying
and interacting with quality indicators in the application.
}
