% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_samlerapport.R
\name{samlerapport_server}
\alias{samlerapport_server}
\title{Shiny module providing server logic for the report tab}
\usage{
samlerapport_server(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\description{
Shiny module providing server logic for the report tab
}
