% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_utsending.R
\name{utsending_server}
\alias{utsending_server}
\title{Shiny module providing server logic for the dispatch tab}
\usage{
utsending_server(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\description{
Shiny module providing server logic for the dispatch tab
}
