% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_veiledning.R
\name{veiledning_server}
\alias{veiledning_server}
\title{Shiny module providing server logic for the intro tab}
\usage{
veiledning_server(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\description{
Shiny module providing server logic for the intro tab
}
