% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeQualind.R
\name{annotations}
\alias{annotations}
\title{Annotations}
\usage{
annotations(var)
}
\value{
data frame with numeric values
}
\description{
Get the right annotations in plot based on desired points of measurement
for each of the quality indicators
}
\details{
If a certain quality indicator is selected by user, the right level must be
indicated in green and in orange. xmax and xmin indicate the level for high
(=green) achievement. xmax_moderate and xmin_moderate indicate the level for
moderate achievement (i.e., orange)
}
\examples{
my_annotations <- annotations("behandlingsplan")

}
