% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegData.R
\name{getBasisData}
\alias{getBasisData}
\title{Provides a dataframe containing data from a registry}
\usage{
getBasisData()
}
\value{
regData data frame
}
\description{
Provides a dataframe containing data from a registry
}
