% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_plots.R
\name{plots_server}
\alias{plots_server}
\title{Shiny module providing server logic for the plot tab}
\usage{
plots_server(id)
}
\arguments{
\item{id}{Character string module namespace}
}
\description{
Shiny module providing server logic for the plot tab
}
